requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
google-auth==1.32.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
importlib-resources==5.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
twilio==6.61.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.4.7
Werkzeug==2.0.1
zipp==3.5.0
APScheduler==3.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
google-auth==2.3.2
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.20.1
idna==3.3
oauth2client==4.1.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
twilio==7.2.0
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.26.7
